(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -4,15 +4,19 @@ import bcrypt from 'bcrypt'
 
 const controller = {}     // Objeto vazio
 
-controller.create = async function (req, res) {
+controller.create = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
 
     // Verifica se existe o campo "password" em "req.body".
     // Caso positivo, geramos o hash da senha antes de enviá-la
     // ao BD
     // (12 na chamada a bcrypt.hash() corresponde ao número de
     // passos de criptografia utilizados no processo)
-    if (req.body.password) {
+    if(req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12)
     }
 
@@ -21,7 +25,7 @@ controller.create = async function (req, res) {
     // HTTP 201: Created
     res.status(201).end()
   }
-  catch (error) {
+  catch(error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -29,18 +33,24 @@ controller.create = async function (req, res) {
   }
 }
 
-controller.retrieveAll = async function (req, res) {
+
+controller.retrieveAll = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+      
     const result = await prisma.user.findMany(
       // Omite o campo "password" do resultado
       // por questão de segurança
-      { omit: { password: true } }
+      { omit: { password: true } } 
     )
 
     // HTTP 200: OK (implícito)
     res.send(result)
   }
-  catch (error) {
+  catch(error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -48,8 +58,17 @@ controller.retrieveAll = async function (req, res) {
   }
 }
 
-controller.retrieveOne = async function (req, res) {
+
+controller.retrieveOne = async function(req, res) {
   try {
+
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if(! (req?.authUser?.is_admin || 
+      Number(req?.authUser?.id) === Number(req.params.id))) 
+      return res.status(403).end()
+      
     const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questão de segurança
@@ -58,11 +77,11 @@ controller.retrieveOne = async function (req, res) {
     })
 
     // Encontrou ~> retorna HTTP 200: OK (implícito)
-    if (result) res.send(result)
+    if(result) res.send(result)
     // Não encontrou ~> retorna HTTP 404: Not Found
     else res.status(404).end()
   }
-  catch (error) {
+  catch(error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -70,15 +89,20 @@ controller.retrieveOne = async function (req, res) {
   }
 }
 
-controller.update = async function (req, res) {
+
+controller.update = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
 
     // Verifica se existe o campo "password" em "req.body".
     // Caso positivo, geramos o hash da senha antes de enviá-la
     // ao BD
     // (12 na chamada a bcrypt.hash() corresponde ao número de
     // passos de criptografia utilizados no processo)
-    if (req.body.password) {
+    if(req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12)
     }
 
@@ -88,11 +112,11 @@ controller.update = async function (req, res) {
     })
 
     // Encontrou e atualizou ~> HTTP 204: No Content
-    if (result) res.status(204).end()
+    if(result) res.status(204).end()
     // Não encontrou (e não atualizou) ~> HTTP 404: Not Found
     else res.status(404).end()
   }
-  catch (error) {
+  catch(error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -101,8 +125,14 @@ controller.update = async function (req, res) {
 }
 
 
-controller.delete = async function (req, res) {
+
+controller.delete = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
@@ -110,8 +140,8 @@ controller.delete = async function (req, res) {
     // Encontrou e excluiu ~> HTTP 204: No Content
     res.status(204).end()
   }
-  catch (error) {
-    if (error?.code === 'P2025') {
+  catch(error) {
+    if(error?.code === 'P2025') {
       // Não encontrou e não excluiu ~> HTTP 404: Not Found
       res.status(404).end()
     }
@@ -124,6 +154,7 @@ controller.delete = async function (req, res) {
     }
   }
 }
+
 
 controller.login = async function (req, res) {
   try {
